feat: moved AssetIcon to ui-entities
