+++ src/renderer/entities/asset/ui/index.ts
@@ -1,5 +1,5 @@
 export { AssetBalance } from './AssetBalance/AssetBalance';
 export { AssetDetails } from './AssetDetails/AssetDetails';
-export { AssetIcon } from './AssetIcon/AssetIcon';
+export { AssetIcon } from '@/shared/ui-entities';
 export { AssetLinks } from './AssetLinks/AssetLinks';
 export { EmptyAssetsState } from './EmptyAssetsState/EmptyAssetsState';
