+++ src/renderer/shared/ui-entities/index.ts
@@ -2,6 +2,7 @@ export { VoteChart } from './VoteChart/VoteChart';
 export { Hash } from './Hash/Hash';
 export { Address } from './Address/Address';
 export { Account } from './Account/Account';
+export { AssetIcon } from './AssetIcon/AssetIcon';
 export { AccountSelectModal } from './AccountSelectModal/AccountSelectModal';
 export { AccountExplorers } from './AccountExplorer/AccountExplorers';
 export { TransactionDetails } from './TransactionDetails/TransactionDetails';
