+++ src/renderer/widgets/AssetTransactionModal/model/__tests__/asset-transaction-model.test.ts
@@ -10,7 +10,11 @@ const mockAsset = {
   name: 'Polkadot',
   precision: 10,
   priceId: 'polkadot',
-  icon: 'https://raw.githubusercontent.com/novasamatech/nova-spektr-utils/main/icons/v1/assets/white/Polkadot_(DOT).svg',
+  icon: {
+    monochrome:
+      'https://raw.githubusercontent.com/novasamatech/nova-spektr-utils/main/icons/v2/assets/monochrome/DOT.svg',
+    colored: 'https://raw.githubusercontent.com/novasamatech/nova-spektr-utils/main/icons/v2/assets/colored/DOT.svg',
+  },
   symbol: 'DOT',
   chains: [
     {
